实现了根据不同的状态, 从 theme 对象中查找相应的 style

+++ guilink/core/src/main/java/com/bitwormhole/guilink/graphics/Stroke.java
@@ -2,6 +2,6 @@ package com.bitwormhole.guilink.graphics;
 
 public class Stroke {
 
-    public int width;
+    public float width;
 
 }
